# examples/crud/openapi_out.yml
# Specification generated by echOpen

openapi: 3.1.0
jsonSchemaDialect: https://spec.openapis.org/oas/3.1/dialect/base
info:
    title: Responses
    version: 1.0.0
    description: Example to show a CRUD-like API, based on Todo lists.
tags:
    - name: todo
      description: Todo endpoints
paths:
    /todo:
        get:
            operationId: getTodo
            description: Get all todos for the user
            tags:
                - todo
            parameters:
                - name: search
                  in: query
                  style: form
                  schema:
                    type: string
                - name: completed
                  in: query
                  style: form
                  schema:
                    type: boolean
                - name: tags
                  in: query
                  style: form
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Todo'
                default:
                    $ref: '#/components/responses/UnexpectedErrorResponse'
        post:
            operationId: postTodo
            description: Create a new Todo
            tags:
                - todo
            requestBody:
                description: New Todo
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NewTodo'
            responses:
                "201":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                "400":
                    $ref: '#/components/responses/BadRequestResponse'
                default:
                    $ref: '#/components/responses/UnexpectedErrorResponse'
    /todo/{id}:
        get:
            operationId: getTodoById
            description: Get todo by ID
            tags:
                - todo
            parameters:
                - name: id
                  in: path
                  description: Todo ID
                  required: true
                  schema:
                    type: string
                    format: uuid
                  examples:
                    - value: 11c7810d-6627-497a-91e9-e3dc4812ce30
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                default:
                    $ref: '#/components/responses/UnexpectedErrorResponse'
        delete:
            operationId: deleteTodoById
            description: Delete a todo
            tags:
                - todo
            parameters:
                - name: id
                  in: path
                  description: Todo ID
                  required: true
                  schema:
                    type: string
                    format: uuid
                  examples:
                    - value: 11c7810d-6627-497a-91e9-e3dc4812ce30
            responses:
                "204":
                    description: Successful response
                "400":
                    $ref: '#/components/responses/BadRequestResponse'
                default:
                    $ref: '#/components/responses/UnexpectedErrorResponse'
        patch:
            operationId: patchTodoById
            description: Update a todo
            tags:
                - todo
            parameters:
                - name: id
                  in: path
                  description: Todo ID
                  required: true
                  schema:
                    type: string
                    format: uuid
                  examples:
                    - value: 11c7810d-6627-497a-91e9-e3dc4812ce30
            requestBody:
                description: Updated Todo
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTodo'
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                "400":
                    $ref: '#/components/responses/BadRequestResponse'
                default:
                    $ref: '#/components/responses/UnexpectedErrorResponse'
components:
    schemas:
        Error:
            type: object
            properties:
                message:
                    type: string
        NewTodo:
            type: object
            required:
                - description
            properties:
                description:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
        Todo:
            type: object
            required:
                - id
                - created
                - updated
            properties:
                completed:
                    type: boolean
                created:
                    type: string
                    format: date-time
                description:
                    type: string
                id:
                    type: string
                    format: uuid
                tags:
                    type: array
                    items:
                        type: string
                updated:
                    type: string
                    format: date-time
        UpdateTodo:
            type: object
            properties:
                completed:
                    type: boolean
                description:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
    responses:
        BadRequestResponse:
            description: Bad request
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        NotFoundResponse:
            description: Not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        UnexpectedErrorResponse:
            description: Unexpected error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'